fix(contracts): break src→tests import — relocate _schema to src/contracts/
PR A pass-2 fresh-eyes review flagged src/contract_diff.py:54 imported
from tests.trust.contracts._schema — a src→tests violation introduced
by PR #8390. Production code shouldn't require tests/ on sys.path
(the deployed container ships src/ only). Same class of bug we fixed
for the Fakes in PR A (factories relocated to src/mockworld/fakes/_factories.py).

Relocated: tests/trust/contracts/_schema.py → src/contracts/_schema.py
(Cassette, CassetteInput, CassetteOutput, NORMALIZERS, apply_normalizers,
load_cassette, dump_cassette).

Back-compat shim left at tests/trust/contracts/_schema.py re-exports
the moved symbols so the eight existing 'from tests.trust.contracts._schema'
call sites in tests/ keep working.

src/contract_diff.py now imports from contracts._schema. Verified:
contract_diff + contract_refresh_loop resolve under PYTHONPATH=src
alone (no tests/ on sys.path) — matches the Docker import path.

Stale docstring references to tests/trust/contracts/_schema.py in
src/contract_recording.py and src/fake_coverage_auditor_loop.py
updated to point at the new src/contracts/_schema.py.

Pyright fix while moving: the src/ tier is stricter than tests/, so the
recorded_at validator's hasattr-based isoformat call had to be replaced
with isinstance(v, datetime|date) — pyright doesn't narrow object via
hasattr.

docs/arch/generated/modules.md regenerated to add the new src.contracts
package node and the src→src.contracts edge (1 import). Other arch
artifacts updated only by the regen footer SHA/timestamp.

Skip-ADR: Internal refactor breaking a src→tests dependency. No
behaviors codified by ADRs change.
